Create hosting view controller for profile editing destination views
